update code cmm013

+++ nts.uk/web/nts.uk.com.web/src/main/webapp/view/cmm/013/a/cmm013.a.vm.js
@@ -36,7 +36,7 @@ var cmm013;
                     self.currentCode = ko.observable();
                     self.clickChange = ko.observable(false);
                     self.columns = ko.observableArray([
-                        { headerText: 'コード', key: 'jobCode', width: 80 },
+                        { headerText: 'コード', key: 'jobCode', width: 90 },
                         { headerText: '名称', key: 'jobName', width: 100 }
                     ]);
                     self.roundingRules = ko.observableArray([
@@ -50,9 +50,12 @@ var cmm013;
                         new BoxModel(2, 'ロール毎に設定')
                     ]);
                     self.selectedId = ko.observable(0);
+                    if (self.selectedId() == 0) {
+                        $('#lst_003').addClass('disableClass');
+                    }
                     self.enable = ko.observable(true);
                     self.notAlert = ko.observable(true);
-                    self.dirty = new nts.uk.ui.DirtyChecker(self.listbox);
+                    self.dirty = new nts.uk.ui.DirtyChecker(self.dataSource);
                     $("#list-box").click(function (evt, ui) {
                         self.clickChange(true);
                     });
@@ -61,6 +64,13 @@ var cmm013;
                         self.oldStartDate(self.itemHist().startDate);
                         self.oldEndDate(self.itemHist().endDate);
                         self.srtDateLast(self.listbox()[0].startDate);
+                        if (codeChanged == null) {
+                            return;
+                        }
+                        if (!self.notAlert()) {
+                            self.notAlert(true);
+                            return;
+                        }
                         var chkCopy = nts.uk.ui.windows.getShared('cmm013Copy');
                         if (codeChanged === '1' && chkCopy) {
                             self.inp_002_enable(true);
@@ -107,21 +117,38 @@ var cmm013;
                         self.selectedId(self.currentItem().presenceCheckScopeSet);
                         self.inp_002_enable(false);
                         self.createdMode(false);
-                        a.service.getAllJobTitleAuth(self.currentItem().historyId, self.currentItem().jobCode).done(function (jTref) {
-                            if (jTref.length === 0) {
-                                $('.trLst003').hide();
-                            }
-                            else {
-                                $('.trLst003').show();
-                                self.dataRef([]);
-                                _.map(jTref, function (item) {
-                                    var tmp = new model.GetAuth(item.jobCode, item.authCode, item.authName, item.referenceSettings);
-                                    self.dataRef.push(tmp);
-                                    return tmp;
+                        a.service.findByUseKt().done(function (res) {
+                            if (res.use_Kt_Set === 1) {
+                                a.service.getAllJobTitleAuth(self.currentItem().historyId, self.currentItem().jobCode).done(function (jTref) {
+                                    if (jTref.length === 0) {
+                                        $('.trLst003').hide();
+                                    }
+                                    else {
+                                        $('.trLst003').show();
+                                        self.dataRef([]);
+                                        _.map(jTref, function (item) {
+                                            var tmp = new model.GetAuth(item.jobCode, item.authCode, item.authName, item.referenceSettings);
+                                            self.dataRef.push(tmp);
+                                            return tmp;
+                                        });
+                                        self.dataRefNew(self.dataRef());
+                                        self.createdMode(true);
+                                    }
                                 });
-                                self.dataRefNew(self.dataRef());
-                                self.createdMode(true);
                             }
+                            self.selectedId.subscribe(function (codeChanged) {
+                                $('#lst_003').removeClass('disableClass');
+                                if (codeChanged == 0) {
+                                    $('#lst_003').show();
+                                    $('#lst_003').addClass('disableClass');
+                                }
+                                else if (codeChanged == 1) {
+                                    $('#lst_003').hide();
+                                }
+                                else {
+                                    $('#lst_003').show();
+                                }
+                            });
                         });
                     }
                 };
@@ -224,7 +251,7 @@ var cmm013;
                     }
                     if (self.inp_003_name() === "" || self.inp_003_name() === null) {
                         nts.uk.ui.dialog.alert("名称が入力されていません");
-                        $('inp_003').focus();
+                        $('#inp_003').focus();
                         return false;
                     }
                     return true;
@@ -257,25 +284,12 @@ var cmm013;
                     if (self.checkChangeData() == false || self.checkChangeData() === undefined) {
                         if (self.dirty.isDirty()) {
                             nts.uk.ui.dialog.confirm("変更された内容が登録されていません。\r\n よろしいですか。").ifYes(function () {
-                                self.inp_002_enable(true);
-                                self.inp_002_code("");
-                                self.inp_003_name("");
-                                self.selectedId(1);
-                                self.inp_005_memo("");
-                                self.currentCode(null);
-                                $("#inp_002").focus();
+                                self.clearInit();
                             }).ifNo(function () {
                             });
                         }
                         else {
-                            self.inp_002_enable(true);
-                            self.inp_002_code("");
-                            self.inp_003_name("");
-                            self.dataRef([]);
-                            self.selectedId(1);
-                            self.inp_005_memo("");
-                            self.currentCode(null);
-                            $("#inp_002").focus();
+                            self.clearInit();
                         }
                     }
                 };
@@ -389,15 +403,19 @@ var cmm013;
                         var dfd = $.Deferred();
                         var item = new model.DeleteJobTitle(self.selectedCode(), self.currentItem().jobCode);
                         self.index_of_itemDelete = self.dataSource().indexOf(self.currentItem());
-                        nts.uk.ui.dialog.confirm("データを削除します。\r\nよろしいですか？").ifYes(function () {
-                            a.service.deleteJobTitle(item).done(function (res) {
-                                self.dirty.reset();
-                                self.getPositionList_afterDelete();
-                            }).fail(function (res) {
-                                dfd.reject(res);
+                        if (self.dataSource().length == 1) {
+                            nts.uk.ui.dialog.confirm("選択している履歴の職位が1件のみのため、\r\n履歴の編集ボタンから履歴削除を行ってください。");
+                        }
+                        else {
+                            nts.uk.ui.dialog.confirm("データを削除します。\r\nよろしいですか？").ifYes(function () {
+                                a.service.deleteJobTitle(item).done(function (res) {
+                                    self.getPositionList_afterDelete();
+                                }).fail(function (res) {
+                                    dfd.reject(res);
+                                });
+                            }).ifNo(function () {
                             });
-                        }).ifNo(function () {
-                        });
+                        }
                     }
                 };
                 ScreenModel.prototype.getPositionList_afterDelete = function () {
@@ -547,6 +565,16 @@ var cmm013;
                     return refJob;
                 }());
                 model.refJob = refJob;
+                var InputField = (function () {
+                    function InputField(position, enable) {
+                        this.inp_002_code = ko.observable(position.jobCode);
+                        this.inp_003_name = ko.observable(position.jobName);
+                        this.inp_004_notes = ko.observable(position.memo);
+                        this.inp_002_enable = ko.observable(enable);
+                    }
+                    return InputField;
+                }());
+                model.InputField = InputField;
             })(model = viewmodel.model || (viewmodel.model = {}));
         })(viewmodel = a.viewmodel || (a.viewmodel = {}));
     })(a = cmm013.a || (cmm013.a = {}));
